+++ geopandas/geodataframe.py
@@ -125,6 +125,9 @@ class GeoDataFrame(GeoPandasBase, DataFrame):
 
     _metadata = ["_geometry_column_name"]
 
+    _internal_names = DataFrame._internal_names + ["geometry"]
+    _internal_names_set = set(_internal_names)
+
     _geometry_column_name = DEFAULT_GEO_COLUMN_NAME
 
     def __init__(self, data=None, *args, geometry=None, crs=None, **kwargs):