+++ CHANGELOG.md
@@ -14,6 +14,8 @@ Deprecations and compatibility notes:
 - resolve ``matplotlib.cm`` warning in ``.explore()`` (#2596)
 
 Bug fixes:
+- Accessing `gdf.geometry` where the active geometry column is missing, and a column named `"geometry"` is present 
+  will now raise an `AttributeError`, rather than returning `gdf["geometry"]` (#2575) 
 
 Notes on (optional) dependencies: