+++ CHANGELOG.md
@@ -14,6 +14,8 @@ New features and improvements:
 Deprecations and compatibility notes:
 
 Bug fixes:
+- Accessing `gdf.geometry` where the active geometry column is missing, and a column named `"geometry"` is present 
+  will now raise an `AttributeError`, rather than returning `gdf["geometry"]` (#2575) 
 
 Notes on (optional) dependencies: